+++ test/feature_test.dart
@@ -1,4 +1,5 @@
 import 'package:bdd_widget_test/src/feature_file.dart';
+import 'package:bdd_widget_test/src/generator_options.dart';
 import 'package:test/test.dart';
 
 import 'util/testing_data.dart';
@@ -105,6 +106,36 @@ void main() {
       package: 'test',
       input: bigFeatureFile,
     );
+    expect(feature.dartContent, expectedFeatureDart);
+  });
+  test('custom headers replace default imports in feature file', () async {
+    const expectedFeatureDart = '''
+${expectedComment}import 'package:patrol/patrol.dart';
+// Import flutter_test for compatibility
+import 'package:flutter_test/flutter_test.dart';
+
+import './step/the_app_is_running.dart';
+
+void main() {
+  group(\'\'\'Testing feature\'\'\', () {
+    testWidgets(\'\'\'Testing scenario\'\'\', (tester) async {
+      await theAppIsRunning(tester);
+    });
+  });
+}
+''';
+
+    final feature = FeatureFile(
+        featureDir: 'test.feature',
+        package: 'test',
+        input: minimalFeatureFile,
+        generatorOptions: const GeneratorOptions(
+          customHeaders: [
+            "import 'package:patrol/patrol.dart';",
+            '// Import flutter_test for compatibility',
+            "import 'package:flutter_test/flutter_test.dart';",
+          ],
+        ));
     expect(feature.dartContent, expectedFeatureDart);
   });
 }